requirements.txt
# Core dependencies
faker==20.0.0
PyQt6==6.6.0
python-dotenv==1.0.0
requests==2.32.4
django>=4.2.0

# Web frameworks
flask==3.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database drivers (optional - install as needed)
psycopg2-binary>=2.9.9  # PostgreSQL
mysql-connector-python>=8.2.0  # MySQL
pymongo>=4.6.1  # MongoDB

# Additional utilities
pyyaml>=6.0.1  # Configuration files
aiofiles>=23.2.1  # Async file operations
